refactor: Station DB 예외 처리 추가

+++ src/main/java/wooteco/subway/station/StationService.java
@@ -14,8 +14,15 @@ public class StationService {
     }
 
     public StationResponse save(Station station) {
+        if (stationNameExists(station.getName())) {
+            throw new IllegalArgumentException("역 이름이 이미 존재합니다.");
+        }
         Station newStation = stationDao.save(station);
         return new StationResponse(newStation.getId(), newStation.getName());
+    }
+
+    private boolean stationNameExists(String name) {
+        return stationDao.count(name) > 0;
     }
 
     public List<StationResponse> findAll() {
@@ -26,6 +33,13 @@ public class StationService {
     }
 
     public void delete(Long id) {
+        if (stationIdNotExists(id)) {
+            throw new IllegalArgumentException("역 ID가 존재하지 않습니다.");
+        }
         stationDao.delete(id);
+    }
+
+    private boolean stationIdNotExists(Long id) {
+        return stationDao.count(id) == 0;
     }
 }
